added interface for the response
