+++ pkg/apiCalls/prepare.go
@@ -89,6 +89,14 @@ func EncodeGraphQlBody(query string, variables map[string]interface{}) (io.Reade
 	return bytes.NewReader(jsonData), nil
 }
 
+// structure of the result to print in the console as the std output
+type CustomResponse struct {
+	Body           interface{} `json:"Body"`
+	ResponseStatus string      `json:"Response Status"`
+}
+
+// Takes in http response, adds response status code,
+// and returns CustomResponse type string for the StandardCall function below
 func processResponse(response *http.Response) string {
 	jsonBody, err := io.ReadAll(response.Body)
 	if err != nil {
@@ -96,15 +104,15 @@ func processResponse(response *http.Response) string {
 	}
 	defer response.Body.Close()
 
-	responseData := map[string]interface{}{
-		"Response Status": response.Status,
+	responseData := CustomResponse{
+		ResponseStatus: response.Status,
 	}
 	var parsedBody interface{}
 	if err := json.Unmarshal(jsonBody, &parsedBody); err == nil {
-		responseData["Body"] = parsedBody
+		responseData.Body = parsedBody
 	} else {
 		// If the body isn't valid JSON, include it as a string
-		responseData["Body"] = string(jsonBody)
+		responseData.Body = string(jsonBody)
 	}
 
 	finalJSON, err := json.MarshalIndent(responseData, "", "  ")
